refactor: Replace useSelector with useAppSelector
