+++ src/components/_taskbar/Calendar/components/YearsList/YearsList.tsx
@@ -2,9 +2,9 @@ import { FC } from 'react';
 
 import useFillYears from 'hooks/useFillYears';
 
+import { useAppDispatch } from 'store/hooks';
 import { calendarActions } from 'store/slices/calendar.slice';
 import { currentDateActions } from 'store/slices/currentDate.slice';
-import { useAppDispatch } from 'store/store';
 
 import { YearType } from 'types/components/calendar/yearType.type';
