+++ src/components/_taskbar/PlansDate/PlansDate.tsx
@@ -1,16 +1,15 @@
 import { FC } from 'react';
-import { useSelector } from 'react-redux';
 
 import PlansClock from './components/PlansClock/PlansClock';
 
-import { RootState } from 'types/store/store.type';
+import { useAppSelector } from 'store/hooks';
 
 import formatCurrentDate from 'utils/formatCurrentDate';
 
 import classes from './PlansDate.module.css';
 
 const PlansDate: FC = () => {
-  const today = useSelector((state: RootState) => state.currentDate.today);
+  const today = useAppSelector(state => state.currentDate.today);
 
   const { currentDateNamesFormat } = formatCurrentDate(today);
