+++ src/components/_taskbar/PlansDate/components/PlansClock/PlansClock.tsx
@@ -1,12 +1,11 @@
 import { FC } from 'react';
-import { useSelector } from 'react-redux';
 
-import { RootState } from 'types/store/store.type';
+import { useAppSelector } from 'store/hooks';
 
 import classes from './PlansClock.module.css';
 
 const PlansClock: FC = () => {
-  const time = useSelector((state: RootState) => state.clock.time);
+  const time = useAppSelector(state => state.clock.time);
 
   return <div className={classes.root}>{time}</div>;
 };
