+++ src/components/_taskbar/TimeAndDate/TimeAndDate.tsx
@@ -1,20 +1,17 @@
 import { FC, memo } from 'react';
-import { useSelector } from 'react-redux';
 
 import CurrentDate from './components/CurrentDate/CurrentDate';
 import CurrentTime from './components/CurrentTime/CurrentTime';
 
+import { useAppDispatch, useAppSelector } from 'store/hooks';
 import { plansActions } from 'store/slices/plans.slice';
-import { useAppDispatch } from 'store/store';
-
-import { RootState } from 'types/store/store.type';
 
 import formatCurrentDate from 'utils/formatCurrentDate';
 
 import classes from './TimeAndDate.module.css';
 
 const TimeAndDate: FC = () => {
-  const today = useSelector((state: RootState) => state.currentDate.today);
+  const today = useAppSelector(state => state.currentDate.today);
   const dispatch = useAppDispatch();
 
   const { currentDateNamesFormat } = formatCurrentDate(today);
