+++ src/components/_taskbar/TimeAndDate/components/CurrentTime/CurrentTime.tsx
@@ -1,12 +1,11 @@
 import { FC } from 'react';
-import { useSelector } from 'react-redux';
 
-import { RootState } from 'types/store/store.type';
+import { useAppSelector } from 'store/hooks';
 
 import classes from './CurrentTime.module.css';
 
 const CurrentTime: FC = () => {
-  const time = useSelector((state: RootState) => state.clock.time);
+  const time = useAppSelector(state => state.clock.time);
   return <div className={classes.root}>{`${time?.substring(0, 5)}`}</div>;
 };
