+++ src/components/_taskbar/ToggleEventsVisibility/ToggleEventsVisibility.tsx
@@ -1,10 +1,7 @@
 import { FC } from 'react';
-import { useSelector } from 'react-redux';
 
+import { useAppDispatch, useAppSelector } from 'store/hooks';
 import { plansActions } from 'store/slices/plans.slice';
-import { useAppDispatch } from 'store/store';
-
-import { RootState } from 'types/store/store.type';
 
 import { ReactComponent as ArrowDownIcon } from 'assets/icons/arrow_down.svg';
 import { ReactComponent as ArrowUpIcon } from 'assets/icons/arrow_up.svg';
@@ -12,7 +9,7 @@ import { ReactComponent as ArrowUpIcon } from 'assets/icons/arrow_up.svg';
 import classes from './ToggleEventsVisibility.module.css';
 
 const ToggleEventsVisibility: FC = () => {
-  const isEventsVisible = useSelector((state: RootState) => state.plans.isEventsVisible);
+  const isEventsVisible = useAppSelector(state => state.plans.isEventsVisible);
   const dispatch = useAppDispatch();
 
   const handleToggleEventsVisibility = () => {
